use the dto in the controller instead of the entity.

+++ src/main/java/com/example/SpaceshipW2M/repository/SpaceshipRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import com.example.SpaceshipW2M.entity.SpaceshipEntity;
 
 @Repository
-public interface SpaceshipRepository extends JpaRepository<SpaceshipEntity, Integer> {
+public interface SpaceshipRepository extends JpaRepository<SpaceshipEntity, Long> {
 
 	List<SpaceshipEntity> findBySpaceshipNameContainingIgnoreCase(String spaceshipName);
